pyproject.toml
[project]
name = "pydantic-project"
version = "0.1.0"
description = "Project with Pydantic and FastAPI, from a Udemy course"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "upstash-redis>=1.4.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "watchfiles>=1.0.5",
]
